first pass at the code for this
probably will need several more bugfix commits because I have to test it not on my laptop
