+++ setup.py
@@ -12,8 +12,8 @@ setupkwargs = dict(
         'console_scripts' : ['UploadDataFile=openmsipython.data_file_io.upload_data_file:main',
                              'DataFileUploadDirectory=openmsipython.data_file_io.data_file_upload_directory:main',
                              'DataFileDownloadDirectory=openmsipython.data_file_io.data_file_download_directory:main',
-                             'InstallService=openmsipython.services.service_manager:install',
-                             'ManageService=openmsipython.services.service_manager:manage',
+                             'InstallService=openmsipython.services.install_service:main',
+                             'ManageService=openmsipython.services.manage_service:main',
                              'LecroyFileUploadDirectory=openmsipython.pdv.lecroy_file_upload_directory:main',
                              'PDVPlotMaker=openmsipython.pdv.pdv_plot_maker:main',
                              'ProvisionNode=openmsipython.utilities.provision_wrapper:main',
